Change type of variable to avoid compiler warning for x64 architecture
git-svn-id: svn://svn.minisip.org/minisip/trunk@2658 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmsip/source/headers/SipHeaderVia.cxx
@@ -53,7 +53,7 @@ const string sipHeaderValueViaTypeStr = "Via";
 SipHeaderValueVia::SipHeaderValueVia(const string &build_from)
 		: SipHeaderValue(SIP_HEADER_TYPE_VIA,sipHeaderValueViaTypeStr)
 {
-	unsigned i=0;
+	size_t i=0;
 	size_t pos=0;
 
 	ip="";
